feat(web-client): set up initial effects for non-electron env
