+++ package.json
@@ -146,6 +146,7 @@
     "eslint": "^7.6.0",
     "eslint-plugin-import": "^2.20.0",
     "fork-ts-checker-webpack-plugin": "^5.0.14",
+    "html-webpack-inline-source-plugin": "@1.0.0-beta.2",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^26.6.3",
     "jest-canvas-mock": "^2.3.1",
@@ -201,6 +202,7 @@
     "ipfs-core": "^0.14.2",
     "ipfs-http-client": "^56.0.1",
     "ipfs-repo": "^14.0.1",
+    "is-electron": "^2.2.1",
     "isomorphic-ws": "^4.0.1",
     "key-did-provider-ed25519": "^1.1.0",
     "key-did-resolver": "^1.4.4",
